Consolidate site section logic
Summary: I'm going to need the site sections in the about page. As a
result, I need to move the logic for fetching titles, URLs, and the
sections themselves out of App.js. Also this fixes an issue where the
default page's info was rendering because I accidentally removed the
exact path matching for it

Test Plan: Reload page and view text. Click around

Author: sumeetv

Reviewers: