Add new CSS palette and restructure js/css files
Summary: Moved css and js files to their own folders so they're easier
to keep track of. Also added a new palette made based on the material-ui
online helper. Not sure if this is finalized, but I can use the same
generator if I want to change the color scheme without having to
refactor things

Test Plan: Load site and each section

Author: sumeetv

Reviewers: