Start implementing new design
Summary: Add a new paper wrapper for the right-side content and
update CSS to allow for dynamic resizing of the window. This is
the first step to getting to the new design. For mobile, I'll
add some rules to render the right side content below the intro
on the left. For now, it's just a desktop view

Test Plan: npm start; resize and click around